requirements.txt
fastapi==0.95.2
uvicorn==0.31.1
sqlalchemy==1.4.51
pydantic==1.10.13
mitmproxy==9.0.1
pytest>=7.4.0
allure-pytest>=2.15.0
requests>=2.31.0
motor>=3.7.0
aiofiles>=24.1.0
jinja2>=3.0.0
namedlist>=1.7.0
python-multipart>=0.0.10